Changed return type to Integer for movie score

+++ src/main/java/com/crejo/moviereviews/model/Review.java
@@ -3,11 +3,11 @@ package com.crejo.moviereviews.model;
 import com.crejo.moviereviews.service.user.UserType;
 
 public class Review {
-    private Double score;
+    private Integer score;
     private User user;
     private UserType userTypeAtReview;
 
-    public Review(User user, Double score, UserType userTypeAtReview) {
+    public Review(User user, Integer score, UserType userTypeAtReview) {
         this.user = user;
         this.score = score;
         this.userTypeAtReview = userTypeAtReview;
@@ -21,11 +21,11 @@ public class Review {
         this.userTypeAtReview = userTypeAtReview;
     }
 
-    public Double getScore() {
+    public Integer getScore() {
         return score;
     }
 
-    public void setScore(Double score) {
+    public void setScore(Integer score) {
         this.score = score;
     }
